+++ Source/SwiftLintFramework/Rules/FirebaseDynamicLinksSchemeURLRule.swift
@@ -2,7 +2,7 @@
 //  FirebaseDynamicLinksSchemeURLRule.swift
 //  SwiftLint
 //
-//  Created by Ibrahim Ulukaya on 3/8/17.
+//  Created by Ibrahim Ulukaya (Google Inc.) on 3/8/17.
 //  Copyright © 2017 Realm. All rights reserved.
 //
 
@@ -22,7 +22,7 @@ public struct FirebaseDynamicLinksSchemeURLRule: ConfigurationProviderRule, OptI
             "class AppDelegate: UIResponder, UIApplicationDelegate {\n" +
             "  func application(_ application: UIApplication," +
             "      didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {\n" +
-            "    FIROptions.default().deepLinkURLScheme = self.customURLScheme\n" +
+            "    FirebaseOptions.defaultOptions()?.deepLinkURLScheme = self.customURLScheme\n" +
             "    return true \n" +
             "  }\n" +
             "}"
@@ -54,7 +54,7 @@ public struct FirebaseDynamicLinksSchemeURLRule: ConfigurationProviderRule, OptI
                     }
                 case .three:
                     for call in method.substructure where call.kind == SwiftExpressionKind.call.rawValue &&
-                        call.name! == "FIROptions.default" {
+                        call.name! == "FirebaseOptions.defaultOptions" {
                             return []
                     }
                 default: break